pyproject.toml
[project]
name = "sensei-doro"
version = "0.1.0"
description = "Sensei Doro"
authors = [{ name = "Shroominic", email = "contact@shroominic.com" }]
dependencies = [
    "python-dotenv>=1.0.1",
    "discord-py>=2.4.0",
    "discord-ext-pages>=0.0.1",
    "discord-ext-menus>=1.1",
    "watchfiles>=1.0.4",
]
readme = "README.md"
requires-python = ">= 3.11"
license = { text = "MIT" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = ["mypy>=1.15.0", "ruff>=0.9.6"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/sensei_doro"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
exclude = ["gptassistant/bot/prompts.py"]
select = ["E", "F", "I"]

[tool.mypy]
strict = true
python_version = "3.11"
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
# warn_untyped_raises = true
